test tslibs.timezones
